add stays to reservations panel
