+++ src/core/types/index.ts
@@ -8,6 +8,7 @@ export * from "./traveler";
 export * from "./travelerState";
 export * from "./trip";
 export * from "./tripScript";
+export * from "./tripStay";
 export * from "./tripPayment";
 export * from "./tripPending";
 export * from "./user";
