+++ src/features/trips/ReservationsPanel/reservations-panel.styles.scss
@@ -39,5 +39,27 @@
         }
       }
     }
+
+    &__stay {
+      width: 100%;
+      columns: 1;
+      gap: 8px;
+      align-items: center;
+
+      &__piece {
+        width: 100%;
+        display: block;
+        padding: 6px 0;
+
+        &__icon {
+          text-align: left;
+          padding-right: 12px;
+        }
+
+        &__text {
+          display: inline;
+        }
+      }
+    }
   }
 }
