+++ src/services/api/stays/index.ts
@@ -1,3 +1,3 @@
-import { getStayByTripId } from "./by-trip";
+import { getStayByTripId, getAllReservedStaysByTripId } from "./by-trip";
 
-export const StaysApiService = { getByTripId: getStayByTripId };
+export const StaysApiService = { getByTripId: getStayByTripId, getAllReservedByTripId: getAllReservedStaysByTripId };
